Fix deploy docs permissions
The deploy docs action does not seem to be able to get the permissions it needs to do the final deployment of documentation (see the error in the most recent action in CMakePPLang [here](https://github.com/CMakePP/CMakePPLang/actions/runs/3751488080/jobs/6372527490)). This action uses the 'personal_token' input with the 'secrets.GITHUB_TOKEN' variable, which may be causing the issue.